data/requirements.txt
pandas==1.0.3
pyuca==1.2
requests==2.21.0 
camelot-py[cv]